Start RTS porting to compact 64-bit

+++ rts/motoko-rts/src/memory.rs
@@ -35,17 +35,18 @@ pub(crate) const GENERAL_MEMORY_RESERVE: usize = 256 * MB;
 /// This function does not take any `Memory` arguments can be used by the generated code.
 pub trait Memory {
     // General allocator working for all GC variants.
-    unsafe fn alloc_words(&mut self, n: Words<u32>) -> Value;
+    // Must be a multiple of `ADDRESS_ALIGNMENT`.
+    unsafe fn alloc_words(&mut self, n: Words<usize>) -> Value;
 
     // Grow the allocated memory size to at least the address of `ptr`.
-    unsafe fn grow_memory(&mut self, ptr: u64);
+    unsafe fn grow_memory(&mut self, ptr: usize);
 }
 
 /// Allocate a new blob.
 /// Note: After initialization, the post allocation barrier needs to be applied to all mutator objects.
 /// For RTS-internal blobs that can be collected by the next GC run, the post allocation barrier can be omitted.
 #[ic_mem_fn]
-pub unsafe fn alloc_blob<M: Memory>(mem: &mut M, size: Bytes<u32>) -> Value {
+pub unsafe fn alloc_blob<M: Memory>(mem: &mut M, size: Bytes<usize>) -> Value {
     let ptr = mem.alloc_words(size_of::<Blob>() + size.to_words());
     // NB. Cannot use `as_blob` here as we didn't write the header yet
     let blob = ptr.get_ptr() as *mut Blob;
@@ -61,11 +62,11 @@ pub unsafe fn alloc_blob<M: Memory>(mem: &mut M, size: Bytes<u32>) -> Value {
 #[ic_mem_fn]
 pub unsafe fn alloc_array<M: Memory>(mem: &mut M, len: u32) -> Value {
     // Array payload should not be larger than half of the memory
-    if len > MAX_ARRAY_SIZE {
+    if len as usize > MAX_ARRAY_SIZE {
         rts_trap_with("Array allocation too large");
     }
 
-    let skewed_ptr = mem.alloc_words(size_of::<Array>() + Words(len));
+    let skewed_ptr = mem.alloc_words(size_of::<Array>() + Words(len as usize));
 
     let ptr: *mut Array = skewed_ptr.get_ptr() as *mut Array;
     (*ptr).header.tag = TAG_ARRAY;
